Split out capacity rpc tests
